browser: final shared activation function
- Also did a small hack for each entrypoint to provide a region provider
  object since the actual code does not work in browser yet.
- Final shared activation method that is now being used by both activate()
  methods.
