+++ src/extensionShared.ts
@@ -20,24 +20,23 @@ import { openUrl } from './shared/utilities/vsCodeUtils'
 import { activate as activateLogger } from './shared/logger/activation'
 import { initializeComputeRegion } from './shared/extensionUtilities'
 import { activate as activateTelemetry } from './shared/telemetry/activation'
-import { getLogger } from './shared/logger'
 import { DefaultAwsContext } from './shared/awsContext'
 import { Settings } from './shared/settings'
-import { RegionProvider, defaultRegion } from './shared/regions/regionProvider'
 import { DefaultAWSClientBuilder } from './shared/awsClientBuilder'
 import { initialize as initializeAuth } from './auth/activation'
 import { LoginManager } from './auth/deprecated/loginManager'
 import { CredentialsStore } from './auth/credentials/store'
 import { initializeAwsCredentialsStatusBarItem } from './auth/ui/statusBarItem'
+import { RegionProvider } from './shared/regions/regionProvider'
 
 /**
- * This is a temporary activate function that I will be moving pieces in to.
- * It will run code that works in both extension and extensionWeb activations.
+ * Activation/setup code that is shared by the regular (nodejs) extension AND browser-compatible extension.
+ * Most setup code should live here, unless there is a reason not to.
  *
- * Eventually this will be renamed and by executed by both activations at startup
- * as a shared function.
+ * @param getRegionProvider - HACK telemetry requires the region provider but we cannot create it yet in this
+ * "shared" function since it breaks in browser. So for now the caller must provide it.
  */
-export async function testActivate(context: vscode.ExtensionContext) {
+export async function activateShared(context: vscode.ExtensionContext, getRegionProvider: () => RegionProvider) {
     // Setup the logger
     const toolkitOutputChannel = vscode.window.createOutputChannel('AWS Toolkit', { log: true })
     await activateLogger(context, toolkitOutputChannel)
@@ -47,12 +46,17 @@ export async function testActivate(context: vscode.ExtensionContext) {
     globals.awsContext = new DefaultAwsContext()
     globals.sdkClientBuilder = new DefaultAWSClientBuilder(globals.awsContext)
     globals.loginManager = new LoginManager(globals.awsContext, new CredentialsStore())
-    setupGlobalsTempStubs()
 
     // some "initialize" functions
     await initializeComputeRegion()
     initialize(context)
-    initializeManifestPaths(context)
+
+    // order matters here
+    globals.manifestPaths.endpoints = context.asAbsolutePath(join('resources', 'endpoints.json'))
+    globals.manifestPaths.lambdaSampleRequests = context.asAbsolutePath(
+        join('resources', 'vs-lambda-sample-request-manifest.xml')
+    )
+    globals.regionProvider = getRegionProvider()
 
     // telemetry
     await activateTelemetry(context, globals.awsContext, Settings.instance)
@@ -60,56 +64,8 @@ export async function testActivate(context: vscode.ExtensionContext) {
     // auth
     await initializeAuth(context, globals.awsContext, globals.loginManager)
     await initializeAwsCredentialsStatusBarItem(globals.awsContext, context)
-}
 
-/**
- * This function is temporary. In the following commits I will move parts of this
- * in to {@link testActivate} as I confirm that they work in both the browser
- * and node
- */
-export async function browserActivate(context: vscode.ExtensionContext) {
-    try {
-        await testActivate(context)
-
-        registerCommands(context)
-    } catch (error) {
-        const stacktrace = (error as Error).stack?.split('\n')
-        // truncate if the stacktrace is unusually long
-        if (stacktrace !== undefined && stacktrace.length > 40) {
-            stacktrace.length = 40
-        }
-        getLogger().error(`Failed to activate extension`, error)
-        throw error
-    }
-}
-
-/**
- * Since we are still incrementally enabling certain functionality
- * in the browser, certain global variables will not have been set
- * and functionality we enabled will not work.
- *
- * This function sets up the minimum-required stubs for the necessary
- * variables to get things working.
- *
- * If needed we can eventually create the real implementations instead
- * of stubbing.
- */
-function setupGlobalsTempStubs() {
-    // This is required for telemetry to run.
-    // The default region is arbitrary for now.
-    // We didn't create an actual instance since it
-    // will require non-trivial work to get the creation
-    // of the instance in the browser working.
-    globals.regionProvider = {
-        guessDefaultRegion: () => defaultRegion,
-    } as RegionProvider
-}
-
-export function initializeManifestPaths(extensionContext: vscode.ExtensionContext) {
-    globals.manifestPaths.endpoints = extensionContext.asAbsolutePath(join('resources', 'endpoints.json'))
-    globals.manifestPaths.lambdaSampleRequests = extensionContext.asAbsolutePath(
-        join('resources', 'vs-lambda-sample-request-manifest.xml')
-    )
+    registerCommands(context)
 }
 
 /**
