+++ src/extensionWeb.ts
@@ -5,7 +5,9 @@
 
 import * as vscode from 'vscode'
 import { setInBrowser } from './common/browserUtils'
-import { browserActivate } from './extensionShared'
+import { getLogger } from './shared/logger'
+import { activateShared } from './extensionShared'
+import { RegionProvider, defaultRegion } from './shared/regions/regionProvider'
 
 export async function activate(context: vscode.ExtensionContext) {
     setInBrowser(true) // THIS MUST ALWAYS BE FIRST
@@ -14,7 +16,22 @@ export async function activate(context: vscode.ExtensionContext) {
         'AWS Toolkit: Browser Mode Under Development. No features are currently provided'
     )
 
-    await browserActivate(context)
+    try {
+        // IMPORTANT: Any new activation code should be done in the function below unless
+        // it is browser specific activation code.
+        await activateShared(context, () => {
+            return {
+                guessDefaultRegion: () => defaultRegion,
+            } as RegionProvider
+        })
+    } catch (error) {
+        const stacktrace = (error as Error).stack?.split('\n')
+        // truncate if the stacktrace is unusually long
+        if (stacktrace !== undefined && stacktrace.length > 40) {
+            stacktrace.length = 40
+        }
+        getLogger().error(`Failed to activate extension`, error)
+    }
 }
 
 export async function deactivate() {}
